requirements.txt
streamlit
mysql-connector-python
pandas
openpyxlgit remote add origin https://github.com/NAMANNIMBLE1/upcl_backend_script.git